requirements.txt
jinja2>=3.0
Pillow>=9.0
